[SPARK-47534][SQL] Move `o.a.s.variant` to `o.a.s.types.variant`
### What changes were proposed in this pull request?

According to https://github.com/apache/spark/pull/45479#pullrequestreview-1946939461, this PR aims to rename `variant` package and the corresponding test suite like the following.
```
- package org.apache.spark.variant;
+ package org.apache.spark.types.variant;
```

```
$ git diff master --stat
 common/variant/src/main/java/org/apache/spark/{ => types}/variant/Variant.java                                   | 2 +-
 common/variant/src/main/java/org/apache/spark/{ => types}/variant/VariantBuilder.java                            | 4 ++--
 common/variant/src/main/java/org/apache/spark/{ => types}/variant/VariantSizeLimitException.java                 | 2 +-
 common/variant/src/main/java/org/apache/spark/{ => types}/variant/VariantUtil.java                               | 2 +-
 sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/variant/variantExpressions.scala           | 2 +-
 sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/expressions/{ => variant}/VariantExpressionSuite.scala | 6 +++---
 6 files changed, 9 insertions(+), 9 deletions(-)
```

### Why are the changes needed?

To make it clear that `variant` package is related to be a type.

### Does this PR introduce _any_ user-facing change?

No. This package is new in Apache Spark 4.0.0.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #45685 from dongjoon-hyun/SPARK-47534.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
